+++ src/salmon/signal.clj
@@ -8,11 +8,12 @@
   "Handles :salmon/early-validate signal by validating `conf` against the schema
   in the `:salmon/early-schema` entry of the component definition. Does nothing if
   there is no such entry."
-  [{:keys [->validation] ::ds/keys [config system]}]
+  [{::ds/keys [config system]}]
   (let [schema (-> system ::ds/component-def :salmon/early-schema)]
     (when-let [errors (and schema (m/explain schema config))]
-      (->validation {:errors errors
-                     :message (merr/humanize errors)}))))
+      (throw
+        (ex-info (str "Validation failed: " (merr/humanize errors))
+          {:errors errors})))))
 
 (defn- first-line [s]
   (if (string? s)